Decompress funcionando, bugs no trash e na decompactação de imagens

+++ Huffman/lib/inc/decompress.h
@@ -8,13 +8,13 @@
 
 int is_bit_set(unsigned char c, int i);
 
-short get_trash_size(FILE *input);
+int get_trash_size(FILE *input);
 
 short get_tree_size(FILE *input);
 
 TREE* get_hufftree(FILE *input, TREE* tree);
 
-void decompress_file(FILE* input, FILE* output, TREE* tree, short trash_size, short size_tree);
+void decompress_file(FILE* input, FILE* output, TREE* tree, int trash_size, short size_tree);
 
 void decompress(char input_path[], char output_path[]);
